Removed the GameJolt link from the download page for now.
Their web upload system is busted. GameJolt is easily the least
important platform we use, so I'll just leave it as-is and stop
pointing people toward it for now. Will see about contacting GameJolt
about the problem at a later date, but right now it's just not a
high priority.